add execution token cache
